+++ special-pages/pages/history/integration-tests/history.page.js
@@ -1,6 +1,7 @@
 import { perPlatform } from 'injected/integration-test/type-helpers.mjs';
 import { Mocks } from '../../../shared/mocks.js';
 import { expect } from '@playwright/test';
+import { generateSampleData } from '../app/mocks/history.mocks.js';
 
 /**
  * @typedef {import('injected/integration-test/type-helpers.mjs').Build} Build
@@ -135,15 +136,15 @@ export class HistoryTestPage {
         expect(params).toStrictEqual({ query, limit: 150, offset });
     }
 
-    /**
-     * @param {string} linkText
-     */
-    async selectsRange(linkText) {
+    async selectsToday() {
         const { page } = this;
-        await page.getByRole('link', { name: linkText }).click();
+        await page.getByLabel('Show history for today').click();
     }
 
-    async onlyRangeIsShown(s) {}
+    async selectsAll() {
+        const { page } = this;
+        await page.getByLabel('Show all history').click();
+    }
 
     /**
      * @param {string} term
@@ -209,5 +210,125 @@ export class HistoryTestPage {
             url,
             target: 'new-window',
         });
+    }
+
+    /**
+     * @param {import('../types/history.ts').DeleteRangeResponse} resp
+     */
+    async deletesHistoryForToday(resp = { action: 'delete' }) {
+        const { page } = this;
+        await page.getByLabel('Show history for today').hover();
+        if (resp.action === 'delete') {
+            page.on('dialog', (dialog) => dialog.accept());
+        } else {
+            page.on('dialog', (dialog) => dialog.dismiss());
+        }
+        await page.getByLabel('Delete history for today').click();
+        const calls = await this.mocks.waitForCallCount({ method: 'deleteRange', count: 1 });
+        expect(calls[0].payload.params).toStrictEqual({ range: 'today' });
+    }
+
+    /**
+     * @param {import('../types/history.ts').DeleteRangeResponse} resp
+     */
+    async deletesHistoryForYesterday(resp = { action: 'delete' }) {
+        const { page } = this;
+        await page.getByLabel('Show history for yesterday').hover();
+        if (resp.action === 'delete') {
+            page.on('dialog', (dialog) => dialog.accept());
+        } else {
+            page.on('dialog', (dialog) => dialog.dismiss());
+        }
+        await page.getByLabel('Delete history for yesterday').click();
+    }
+
+    async sideBarItemWasRemoved(label) {
+        const { page } = this;
+        await expect(page.getByLabel(label)).not.toBeVisible({ timeout: 1000 });
+    }
+
+    async sidebarHasItem(label) {
+        const { page } = this;
+        await expect(page.getByLabel(label)).toBeVisible({ timeout: 1000 });
+    }
+
+    /**
+     * @param {import('../types/history.ts').DeleteRangeResponse} resp
+     */
+    async deletesAllHistoryFromHeader(resp) {
+        const { page } = this;
+        if (resp.action === 'delete') {
+            page.on('dialog', (dialog) => dialog.accept());
+        } else {
+            page.on('dialog', (dialog) => dialog.dismiss());
+        }
+        await page.getByRole('button', { name: 'Delete All', exact: true }).click();
+        const calls = await this.mocks.waitForCallCount({ method: 'deleteRange', count: 1 });
+        expect(calls[0].payload.params).toStrictEqual({ range: 'all' });
+        await expect(page.getByRole('heading', { level: 2, name: 'Nothing to see here!' })).toBeVisible();
+    }
+
+    /**
+     * @param {import('../types/history.ts').DeleteRangeResponse} resp
+     */
+    async deletesFromSectionTitle(resp) {
+        const { page } = this;
+
+        // Handle dialog interaction based on response action
+        if (resp.action === 'delete') {
+            page.on('dialog', (dialog) => {
+                return dialog.accept();
+            });
+        } else {
+            page.on('dialog', (dialog) => dialog.dismiss());
+        }
+
+        // Hover over the "Today" section and open the menu
+        const title = page.getByRole('list').getByText('Today');
+        await title.hover();
+        await title.getByLabel('Show menu for Today').click();
+
+        // Verify the call to "title_menu" with expected parameters
+        const calls = await this.mocks.waitForCallCount({ method: 'title_menu', count: 1 });
+        expect(calls[0].payload.params).toStrictEqual({ dateRelativeDay: 'Today' });
+
+        // verify the section is gone
+        await expect(title.getByLabel('Show menu for Today')).not.toBeVisible();
+
+        // todo: re-enable this if it's required
+        // await this.sideBarItemWasRemoved('Today');
+    }
+
+    /**
+     * @param {import('../types/history.ts').DeleteRangeResponse} resp
+     */
+    async deletesFromHistoryEntry(resp) {
+        const { page } = this;
+
+        // Handle dialog interaction based on response action
+        if (resp.action === 'delete') {
+            page.on('dialog', (dialog) => {
+                return dialog.accept();
+            });
+        } else {
+            page.on('dialog', (dialog) => dialog.dismiss());
+        }
+        // console.log(data[0].title);
+        const data = generateSampleData({ count: this.entries, offset: 0 });
+        const first = data[0];
+        const row = page.getByText(first.title);
+        await row.hover();
+        await page.locator(`[data-row-menu][value=${data[0].id}]`).click();
+
+        const calls = await this.mocks.waitForCallCount({ method: 'entries_menu', count: 1 });
+        expect(calls[0].payload.params).toStrictEqual({ ids: [data[0].id] });
+    }
+
+    async rightClicksSectionTitle() {
+        const { page } = this;
+        const title = page.getByRole('list').getByText('Today');
+        await title.click({ button: 'right' });
+        const calls = await this.mocks.waitForCallCount({ method: 'title_menu', count: 1 });
+        expect(calls[0].payload.params).toStrictEqual({ dateRelativeDay: 'Today' });
     }
 }